Backend/gatherData/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-core
google-cloud-firestore
requests